requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.8.0
python-decouple==3.8
requests==2.32.5
PyPDF2==3.0.1
python-docx==1.2.0
python-pptx==1.0.2
openai==1.107.2
